+++ AssetsFlow.Client.React/CHANGELOG.md
@@ -9,6 +9,7 @@
 - All sorta of logic not working /refreshing - trade action buttons' state, summary state
 - CompositeFK name mismatch (server issue tbf)
 - Reducing position
+- evalute wasnt updating 
 
 ## [0.9.018-c] - 2025-06-23
 ### Fixed